Add VSC's integrated diff tool, test case ranges

+++ test.sh
@@ -1,8 +1,14 @@
 #!/bin/bash
-for ((i=1; i<=4; i++)) do # Loops through the 4 test cases, might need to change the number depending on how many test cases you have
-    echo "Testing case $i" # prints which test case number it's currently testing 
-
-    # Compiles the java file and runs it with the input file and outputs the result to myout$i.txt(depends on which test case it's currently testing)
-    java -cp bin/ $1 < test/input$i.txt > test/myout$i.txt
-    diff test/myout$i.txt test/output$i.txt
+for ((i=$2; i<=$3; i++)) do
+    if (test -f test/input$i.txt && test -f test/output$i.txt); then
+        java -cp bin/ $1 < test/input$i.txt > test/myout$i.txt
+        if diff test/myout$i.txt test/output$i.txt &>/dev/null; then
+            echo -e "\e[32m[PASS] Test Case $i"
+        else
+            echo -e "\e[31m[FAIL] Test Case $i"
+            code --diff test/myout$i.txt test/output$i.txt
+        fi
+    else
+        echo -e "\e[33mWarning: Failed to locate files for Test Case $i, ensure both input$i.txt and output$i.txt are present in the test directory"
+    fi
 done
